refactoring: grid column 함수 utils로 분리
